Refactor zodHelper file name
Lets keep all the utils in the shared utils folder with a speaking file name.
